fix(chat/video): Refactor imports in chat.py and video_generate.py to improve clarity
